Fixed merge conflicts
