+++ src/test/java/puppy/login/LoginTest.java
@@ -7,10 +7,10 @@ import puppy.base.BaseTest;
 
 
 public class LoginTest extends BaseTest {
-    private final LoginPage loginPage = new LoginPage();
-    private final DashboardPage dashboardPage = new DashboardPage();
     private static final String USERNAME = "admin";
     private static final String PASSWORD = "password";
+    private final LoginPage loginPage = new LoginPage();
+    private final DashboardPage dashboardPage = new DashboardPage();
 
     @Test
     public void loginTest() {
