+++ src/test/java/puppy/login/RegistrationHappyFlowTest.java
@@ -1,15 +1,18 @@
+package puppy.login;
+
 import org.example.pages.DashboardPage;
 import org.example.pages.LoginPage;
 import org.example.pages.RegistrationPage;
 import org.junit.jupiter.api.Test;
+import puppy.base.BaseTest;
 
 public class RegistrationHappyFlowTest extends BaseTest {
 
+    private static final String USERNAME = "Testaccount5";
+    private static final String PASSWORD = "AbobaPuppy";
     private final RegistrationPage registrationPage = new RegistrationPage();
     private final LoginPage loginPage = new LoginPage();
     private final DashboardPage dashboardPage = new DashboardPage();
-    private static final String USERNAME = "Testaccount5";
-    private static final String PASSWORD = "AbobaPuppy";
 
     @Test
     public void registrationHappyFlowTest() {
