+++ src/test/java/puppy/login/WrongLoginTest.java
@@ -5,13 +5,12 @@ import org.junit.jupiter.api.Test;
 import puppy.base.BaseTest;
 
 public class WrongLoginTest extends BaseTest {
-    private final LoginPage loginPage = new LoginPage();
     private static final int RANDOM_USERNAME_LENGHT = 6;
     private static final String PASSWORD = "password";
-
+    private final LoginPage loginPage = new LoginPage();
 
     @Test
-    public void wrongLoginTest () {
+    public void wrongLoginTest() {
         loginPage
                 .navigateToPuppyPage()
                 .enterRandomUsername(RANDOM_USERNAME_LENGHT)
