+++ src/test/java/puppy/login/WrongPasswordTest.java
@@ -5,13 +5,12 @@ import org.junit.jupiter.api.Test;
 import puppy.base.BaseTest;
 
 public class WrongPasswordTest extends BaseTest {
-    private final LoginPage loginPage = new LoginPage();
     private static final String USERNAME = "admin";
     private static final int RANDOM_PASSWORD_LENGHT = 6;
-
+    private final LoginPage loginPage = new LoginPage();
 
     @Test
-    public void wrongPasswordTest () {
+    public void wrongPasswordTest() {
         loginPage
                 .navigateToPuppyPage()
                 .enterUsername(USERNAME)
